requirements.txt
asgiref==3.8.1
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
